Ajustes na saida dos arquivos para criar novos testes

+++ lib/Sort.h
@@ -7,19 +7,19 @@ void print (unsigned long int *v);
 void preencher (unsigned long int *v);
 
 void START (void);
-int verificaSeExiste(unsigned long int v[], int pos, int valor);
+int verificaSeExiste(unsigned long int v[],unsigned long  int pos,unsigned long  int valor);
 void criaVetorSemRepeticao(unsigned long int v[]);
 
 void criaVetorNoventaPorcentoIgual (unsigned long int v[]);
 void verificaQuantidadeDeRepeticoes (unsigned long int v[]);
 
-int crescente (int a, int b);
-int decrescente (int a, int b);
+int crescente (unsigned long int a,unsigned long  int b);
+int decrescente (unsigned long int a,unsigned long  int b);
 
-void HeapSort (unsigned long int vet[], int inicio, int fim, int (*orderby) (int, int));
-void QuickSort (unsigned long int vetor[], int inicio, int fim, int (*orderby) (int, int));
-void RadixSort(unsigned long int vetor[], int inicio, int fim, int (*orderby) (int, int));
-void ShellSort(unsigned long int vetor[], int inicio, int fim, int (*orderby) (int, int));
+void HeapSort (unsigned long int vet[], int inicio, int fim, int (*orderby) (unsigned long int,unsigned long  int));
+void QuickSort (unsigned long int vetor[], int inicio, int fim, int (*orderby) (unsigned long int,unsigned long  int));
+void RadixSort(unsigned long int vetor[], int inicio, int fim, int (*orderby) (unsigned long int,unsigned long  int));
+void ShellSort(unsigned long int vetor[], int inicio, int fim, int (*orderby) (unsigned long int,unsigned long  int));
 
 #include "../src/Sort.cpp"
