+++ lib/TesteTempo.h
@@ -4,10 +4,10 @@
 
 void tempo_gasto (double tempo[][4]);
 
-void Sort_crescente (unsigned long int *v, int N,   double *tempo, int pos, void (*type_sort) (unsigned long int *, int, int, int (*ordem) (int, int)));
-void Sort_decrescente_crescente (unsigned long int *v, int N,  double *tempo, void (*type_sort) (unsigned long int *, int, int, int (*ordem) (int, int)));
-void Sort_crescente_meio_decrescente (unsigned long int *v, int N,  double *tempo, void (*type_sort) (unsigned long int *, int, int, int (*ordem) (int, int)));
-void Sort_decrescente_meio_crescente (unsigned long int *v, int N,  double *tempo, void (*type_sort) (unsigned long int *, int, int, int (*ordem) (int, int)));
+void Sort_crescente (unsigned long int *v, int N,   double *tempo, int pos, void (*type_sort) (unsigned long int *, int, int, int (*ordem) (unsigned long int, unsigned long int)));
+void Sort_decrescente_crescente (unsigned long int *v, int N,  double *tempo, void (*type_sort) (unsigned long int *, int, int, int (*ordem) (unsigned long int,unsigned long  int)));
+void Sort_crescente_meio_decrescente (unsigned long int *v, int N,  double *tempo, void (*type_sort) (unsigned long int *, int, int, int (*ordem) (unsigned long int, unsigned long int)));
+void Sort_decrescente_meio_crescente (unsigned long int *v, int N,  double *tempo, void (*type_sort) (unsigned long int *, int, int, int (*ordem) (unsigned long int,unsigned long  int)));
 
 
 void carregaTamanhoArrayPadrao();
